Moved requester .cpp code to .hpp header
