+++ rosidl_typesupport_opendds_cpp/include/rosidl_typesupport_opendds_cpp/requester_parameters.h
@@ -58,6 +58,16 @@ public:
     return *this;
   }
 
+  RequesterParams& request_topic(const DDS::Topic_var& topic) {
+    request_topic_ = topic;
+    return *this;
+  }
+
+  RequesterParams& reply_topic(const DDS::Topic_var& topic) {
+    reply_topic_ = topic;
+    return *this;
+  }
+
   /*
   GET
   */
@@ -93,6 +103,14 @@ public:
     return reply_topic_name_;
   }
 
+  DDS::Topic_var request_topic() const {
+    return request_topic_;
+  }
+
+  DDS::Topic_var reply_topic() const {
+    return reply_topic_;
+  }
+
 private:
   DDS::DomainParticipant_var participant_;
   DDS::Publisher_var dds_publisher_;
@@ -102,6 +120,8 @@ private:
   std::string service_name_;
   std::string request_topic_name_;
   std::string reply_topic_name_;
+  DDS::Topic_var request_topic_;
+  DDS::Topic_var reply_topic_;
 };
 
 }
